refactor: 파일 구조 변경 및 코드 수정 도메인 폴더 구조 세분화 depth 관련 이름 변경 DocumentList 관련 prop 줄이기
