+++ src/App.js
@@ -1,11 +1,7 @@
-import { putDocument } from "./api/document.js";
+import { getDocuments, putDocument } from "./api/document.js";
 import Layout from "./components/common/Layout.js";
-import DocumentList from "./components/domain/DocumentList.js";
-import DocumentEditor from "./components/domain/DocumentEditor.js";
-import Home from "./components/domain/Home.js";
 import { PATH } from "./constants/path.js";
 import { initRouter, push } from "./utils/route.js";
-import RecurDocumentList from "./components/template/RecurDocumentList.js";
 import {
   TrieDocument,
   addChildDocument,
@@ -16,6 +12,10 @@ import {
 } from "./utils/document.js";
 import NotFound from "./components/common/Notfound.js";
 import { debounce } from "./utils/debounce.js";
+import DocumentList from "./components/domain/document/DocumentList.js";
+import DocumentEditor from "./components/domain/document/DocumentEditor.js";
+import Home from "./components/domain/home/Home.js";
+import RecurDocumentList from "./components/domain/document/template/RecurDocumentList.js";
 
 /**
  * @param {{appElement: Element | null}}
@@ -42,13 +42,11 @@ export default function App({ appElement }) {
 
   this.setState = (nextState) => {
     this.state = nextState;
-
     documentListComponent.render();
   };
 
   this.editorSetState = (nextState) => {
     this.state = nextState;
-
     documentEditorComponent.render();
   };
 
@@ -57,12 +55,11 @@ export default function App({ appElement }) {
   const documentListComponent = new DocumentList({
     parentElement: leftContainerElement,
     renderItemComponent: (parentElement) => {
-      return RecurDocumentList({
+      RecurDocumentList({
         rootDocuments: this.state,
         parentElement,
         childRender: (parentId, newDocument) => {
           const nextState = addChildDocument(parentId, this.state, newDocument);
-
           this.setState(nextState);
         },
         removeRender: (documentId) => {
@@ -75,16 +72,11 @@ export default function App({ appElement }) {
 
           this.setState(newState);
         },
-        count: 1,
+        depthCount: 1,
       });
-    },
-    serverRender: (newState) => {
-      this.setState(newState);
-      insertAllDocument(newState, (title, id) => trie.insert(title, id));
     },
     onAddButtonClick: (newDocument) => {
       const nextState = [...this.state, newDocument];
-
       this.setState(nextState);
     },
   });
@@ -97,11 +89,10 @@ export default function App({ appElement }) {
   const documentEditorComponent = new DocumentEditor({
     parentElement: rightContainerEleement,
     onEditing: (document) => {
-      const { documentId, title } = document;
+      const { documentId, title, isChangeTitle } = document;
 
-      if (document.isChangeTitle) {
+      if (isChangeTitle) {
         const newState = editTitleDocument(documentId, this.state, title);
-
         this.setState(newState);
       }
 
@@ -121,16 +112,18 @@ export default function App({ appElement }) {
 
   initRouter(() => this.route());
 
-  this.init = () => {
-    documentListComponent.getServer();
-
-    layoutComponent.render();
-    documentListComponent.render();
-
-    this.route();
-
+  this.init = async () => {
     appElement.append(wrapperContainer);
     wrapperContainer.append(leftContainerElement, rightContainerEleement);
+
+    layoutComponent.render();
+
+    const newState = await getDocuments();
+    this.setState(newState);
+
+    insertAllDocument(newState, (title, id) => trie.insert(title, id));
+
+    this.route();
   };
 
   this.route = () => {
