+++ src/components/domain/document/DocumentItem.js
@@ -1,4 +1,4 @@
-import Tooltip from "../common/Tooltip.js";
+import Tooltip from "../../common/Tooltip.js";
 
 export default function DocumentItem({
   parentElement,
@@ -6,12 +6,12 @@ export default function DocumentItem({
   onClickChildButton,
   onClickRemoveButton,
   onClickRoute,
-  count,
+  depthCount,
   ...documentData
 }) {
   const containerElement = document.createElement("div");
   containerElement.className = "document-container";
-  containerElement.style.setProperty("--count", count > 10 ? 0 : "20px");
+  containerElement.style.setProperty("--depth", depthCount > 10 ? 0 : "20px");
 
   const tooltipChildAddElement = new Tooltip({ text: "하위 페이지 추가" });
   const tooltipRemoveElement = new Tooltip({ text: "삭제" });
