+++ src/components/domain/document/template/RecurDocumentList.js
@@ -1,20 +1,18 @@
-import { deleteDocument, postDocument } from "../../api/document.js";
-import { PATH } from "../../constants/path.js";
-import { push } from "../../utils/route.js";
-import DocumentItem from "../domain/DocumentItem.js";
+import { deleteDocument, postDocument } from "../../../../api/document.js";
+import { PATH } from "../../../../constants/path.js";
+import { push } from "../../../../utils/route.js";
+import DocumentItem from "../DocumentItem.js";
 
 /**
- * @param {rootDocument} listData
- * @param {parentElement} listElement
  * @description 트리 형태로 문서를 불러오는 함수
  */
 
 export default function RecurDocumentList({
-  rootDocuments,
   parentElement,
+  rootDocuments,
   childRender,
   removeRender,
-  count,
+  depthCount,
 }) {
   rootDocuments.map((rootDocument) =>
     new DocumentItem({
@@ -28,7 +26,7 @@ export default function RecurDocumentList({
                 parentElement: innerParentElement,
                 childRender,
                 removeRender,
-                count: count + 1,
+                depthCount: depthCount + 1,
               }),
       onClickChildButton: async (documentId) => {
         const newDocument = await postDocument({
@@ -45,7 +43,7 @@ export default function RecurDocumentList({
       onClickRoute: async (documentId) => {
         push(`${PATH.DOCUMENTS}/${documentId}`);
       },
-      count,
+      depthCount,
       ...rootDocument,
     }).render()
   );
