+++ src/styles/styles.css
@@ -76,7 +76,7 @@ button {
 .document-container {
   display: flex;
   flex-direction: column;
-  padding-left: var(--count);
+  padding-left: var(--depth);
 }
 
 .toggle-button {
